+++ common.py
@@ -56,6 +56,7 @@ def get_sources(env, is_editor_build):
 
         "engine/*.cpp",
         "engine/gpu/*.cpp",
+        "engine/detail_rendering/*.cpp",
 
         "edition/*.cpp",
         "shaders/*.cpp",
@@ -65,6 +66,7 @@ def get_sources(env, is_editor_build):
         # Utility
 
         "util/*.cpp",
+        "util/containers/*.cpp",
         "util/math/*.cpp",
         "util/noise/fast_noise_lite/*.cpp",
         "util/noise/gd_noise_range.cpp",
@@ -77,6 +79,7 @@ def get_sources(env, is_editor_build):
         "util/godot/classes/array_mesh.cpp",
         "util/godot/classes/concave_polygon_shape_3d.cpp",
         "util/godot/classes/geometry_2d.cpp",
+        "util/godot/classes/geometry_instance_3d.cpp",
         "util/godot/classes/input_event_key.cpp",
         "util/godot/classes/mesh.cpp",
         "util/godot/classes/multimesh.cpp",
@@ -90,12 +93,14 @@ def get_sources(env, is_editor_build):
 
         "util/godot/core/string.cpp",
         "util/godot/core/variant.cpp",
+        "util/godot/core/packed_arrays.cpp",
 
         "util/godot/direct_mesh_instance.cpp",
         "util/godot/direct_multimesh_instance.cpp",
         "util/godot/direct_static_body.cpp",
         "util/godot/shader_material_pool.cpp",
-        "util/godot/funcs.cpp",
+
+        "util/io/*.cpp",
 
         # Thirdparty